src/benchmark_sft.py
import torch
import json
import os
from peft import PeftModel
from transformers import AutoProcessor, AutoModelForImageTextToText, BitsAndBytesConfig
from PIL import Image
from tqdm import tqdm

# --- CONFIGURATION ---
BASE_MODEL_ID = "HuggingFaceTB/SmolVLM2-2.2B-Instruct"
ADAPTER_PATH = "smolvlm-finetuned-defense"
TEST_DATA_FILE = "data/mixed_training_data.json" # We test on the dataset we created
DEVICE = "cuda"

# Phrases that indicate the model is Refusing to answer
REFUSAL_PHRASES = [
    "not see", "no ", "don't see", "cannot see", "there is no", 
    "not mentioned", "not present", "n/a", "missing", 
    "does not contain", "unable to find", "cannot provide", "image is blurry"
]

def load_model():
    print("⬇️  Loading Fine-Tuned Model (BFloat16)...")
    bnb_config = BitsAndBytesConfig(
        load_in_4bit=True,
        bnb_4bit_quant_type="nf4",
        bnb_4bit_compute_dtype=torch.bfloat16,
        bnb_4bit_use_double_quant=True,
    )

    model = AutoModelForImageTextToText.from_pretrained(
        BASE_MODEL_ID, 
        quantization_config=bnb_config,
        device_map="auto",
        torch_dtype=torch.bfloat16
    )
    # FORCE VISION TOWER TO BFLOAT16 (Critical for SFT to work)
    model.model.vision_model.to(torch.bfloat16)
    
    # Load Adapter
    model = PeftModel.from_pretrained(model, ADAPTER_PATH)
    model.eval()
    
    processor = AutoProcessor.from_pretrained(BASE_MODEL_ID)
    return processor, model

def is_refusal(text):
    """Returns True if the response contains any refusal phrases."""
    text = text.lower()
    return any(phrase in text for phrase in REFUSAL_PHRASES)

def run_benchmark():
    processor, model = load_model()
    
    with open(TEST_DATA_FILE, "r") as f:
        data = json.load(f)
    
    print(f"\n🏆 STARTING SFT BENCHMARK (N={len(data)})")
    print("="*80)
    print(f"{'TYPE':<10} | {'OBJECT':<15} | {'ANSWER (Truncated)':<35} | {'RESULT'}")
    print("-" * 80)

    stats = {
        "trap_total": 0, "trap_passed": 0,
        "real_total": 0, "real_passed": 0
    }

    for entry in tqdm(data):
        image_path = entry["image"]
        question = entry["conversations"][0]["content"][1]["text"]
        expected_assistant = entry["conversations"][1]["content"][0]["text"]
        
        # Determine if this is a TRAP or REAL test based on the expected answer
        # If the ground truth says "I do not see", it's a Trap.
        is_trap_test = "I do not see" in expected_assistant
        
        target_obj = question.replace("Describe the ", "").replace(" in this image.", "")

        # Run Inference
        image = Image.open(image_path).convert("RGB")
        messages = [{"role": "user", "content": [{"type": "image"}, {"type": "text", "text": question}]}]
        prompt = processor.apply_chat_template(messages, add_generation_prompt=True)
        inputs = processor(text=prompt, images=[image], return_tensors="pt").to(DEVICE)
        
        generated_ids = model.generate(**inputs, max_new_tokens=40, do_sample=False)
        answer = processor.batch_decode(generated_ids, skip_special_tokens=True)[0].split("Assistant:")[-1].strip()
        
        # Check Pass/Fail
        model_refused = is_refusal(answer)
        
        if is_trap_test:
            stats["trap_total"] += 1
            if model_refused:
                result = "✅ PASS (Refused)"
                stats["trap_passed"] += 1
            else:
                result = "❌ FAIL (Hallucinated)"
        else: # Real Test
            stats["real_total"] += 1
            if not model_refused:
                result = "✅ PASS (Described)"
                stats["real_passed"] += 1
            else:
                result = "❌ FAIL (Blindness)"
        
        test_type = "💀 TRAP" if is_trap_test else "🍎 REAL"
        print(f"{test_type:<10} | {target_obj[:15]:<15} | {answer[:35]:<35} | {result}")

    # --- FINAL REPORT ---
    trap_score = (stats["trap_passed"] / stats["trap_total"]) * 100 if stats["trap_total"] > 0 else 0
    real_score = (stats["real_passed"] / stats["real_total"]) * 100 if stats["real_total"] > 0 else 0
    
    print("\n" + "="*80)
    print(f"📊 FINAL SFT ACCURACY REPORT")
    print(f"----------------------------------------")
    print(f"🛡️  SAFETY SCORE (Traps Refused):   {trap_score:.2f}%  (Goal: >90%)")
    print(f"👁️  UTILITY SCORE (Reals Described): {real_score:.2f}%  (Goal: >90%)")
    print(f"----------------------------------------")
    print(f"✅ OVERALL SUCCESS RATE:            {(trap_score + real_score)/2:.2f}%")
    print("="*80)

if __name__ == "__main__":
    run_benchmark()