requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

# --- Core PyTorch (Keep your specific CUDA version) ---
torch
torchvision
torchaudio

# --- Base VLM Inference ---
transformers>=4.44.0
accelerate>=0.26.0
pillow
scipy
num2words

# --- NEW: Training & QLoRA Stack ---
bitsandbytes>=0.43.0      # Optimized 4-bit quantization
peft>=0.7.0               # LoRA adapters
trl>=0.8.0                # SFTTrainer (The training loop)
datasets                  # To load the JSON data